@@ -1,5 +1,5 @@
 <?php
-namespace dkd\TcBeuser\Utility;
+namespace Dkd\TcBeuser\Utility;
 
 /***************************************************************
 *  Copyright notice
@@ -31,22 +31,39 @@ use TYPO3\CMS\Core\Messaging\FlashMessage;
 use TYPO3\CMS\Core\Messaging\FlashMessageService;
 use TYPO3\CMS\Core\Type\Bitmask\Permission;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Recordlist\RecordList\DatabaseRecordList;   
+use TYPO3\CMS\Recordlist\RecordList\DatabaseRecordList;
 use TYPO3\CMS\Recordlist\RecordList\RecordListHookInterface;
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Database\Query\QueryHelper;
+use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
+use TYPO3\CMS\Core\Database\Query\QueryBuilder;
 
 /**
  * class for listing DB tables in tc_beuser
- * $Id$
  *
  * @author Ingo Renner <ingo.renner@dkd.de>
  * @author Ivan Kartolo <ivan.kartolo@dkd.de>
  */
 class RecordListUtility extends DatabaseRecordList
 {
-
     public $showFields;
     public $userMainGroupOnly = false;
     public $hideDisabledRecords = false;
+
+    /**
+     * Backend users to exclude
+     *
+     * @var string[]
+     */
+    public $excludeBE = [];
+
+    /**
+     * If TRUE, will disable the rendering of clipboard + control panels.
+     *
+     * @var bool
+     */
+    public $dontShowClipControlPanels = false;
+
     /**
      * @var array $disableControls: disable particular features (=icons)
      */
@@ -67,7 +84,6 @@ class RecordListUtility extends DatabaseRecordList
      */
     public function generateList()
     {
-
             // Set page record in header
         $this->pageRecord = BackendUtility::getRecordWSOL('pages', $this->id);
 
@@ -130,37 +146,34 @@ class RecordListUtility extends DatabaseRecordList
     /**
      * Creates the listing of records from a single table
      *
-     * @param string $table : Table name
-     * @param integer $id : Page id
-     * @param string $rowList: List of fields to show in the listing.
-     * Pseudo fields will be added including the record header.
-     * @return string $out: HTML table with the listing for the record.
+     * @param string $table Table name
+     * @param int $id Page id
+     * @param string $rowList List of fields to show in the listing. Pseudo fields will be added including the record header.
+     * @throws \UnexpectedValueException
+     * @return string HTML table with the listing for the record.
      */
     public function getTable($table, $id, $rowList = '')
     {
         $rowListArray = GeneralUtility::trimExplode(',', $rowList, true);
         // if no columns have been specified, show description (if configured)
         if (!empty($GLOBALS['TCA'][$table]['ctrl']['descriptionColumn']) && empty($rowListArray)) {
-            array_push($rowListArray, $GLOBALS['TCA'][$table]['ctrl']['descriptionColumn']);
+            $rowListArray[] = $GLOBALS['TCA'][$table]['ctrl']['descriptionColumn'];
         }
         $backendUser = $this->getBackendUserAuthentication();
         $lang = $this->getLanguageService();
-        $db = $this->getDatabaseConnection();
         // Init
         $addWhere = '';
+        $queryBuilder = $this->getQueryBuilderForTable($table);
         $titleCol = $GLOBALS['TCA'][$table]['ctrl']['label'];
         $thumbsCol = $GLOBALS['TCA'][$table]['ctrl']['thumbnail'];
         $l10nEnabled = $GLOBALS['TCA'][$table]['ctrl']['languageField']
-            && $GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']
-            && !$GLOBALS['TCA'][$table]['ctrl']['transOrigPointerTable'];
+                     && $GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']
+                     && $table !== 'pages_language_overlay';
         $tableCollapsed = (bool)$this->tablesCollapsed[$table];
         // prepare space icon
-        $this->spaceIcon = '<span class="btn btn-default disabled">' .
-            $this->iconFactory->getIcon('empty-empty', Icon::SIZE_SMALL)->render() .
-            '</span>';
-
+        $this->spaceIcon = '<span class="btn btn-default disabled">' . $this->iconFactory->getIcon('empty-empty', Icon::SIZE_SMALL)->render() . '</span>';
         // Cleaning rowlist for duplicates and place the $titleCol as the first column always!
-        $this->fieldArray = array();
+        $this->fieldArray = [];
 
         // ingo.renner@dkd.de
         $this->fieldArray = explode(',', $rowList);
@@ -173,6 +186,10 @@ class RecordListUtility extends DatabaseRecordList
         if ($this->showClipboard) {
             $this->fieldArray[] = '_CLIPBOARD_';
         }
+        // Ref
+        if (!$this->dontShowClipControlPanels) {
+            $this->fieldArray[] = '_REF_';
+        }
         // Path
         if ($this->searchLevels) {
             $this->fieldArray[] = '_PATH_';
@@ -181,11 +198,13 @@ class RecordListUtility extends DatabaseRecordList
         if ($this->localizationView && $l10nEnabled) {
             $this->fieldArray[] = '_LOCALIZATION_';
             $this->fieldArray[] = '_LOCALIZATION_b';
-            $addWhere .= ' AND (
-				' . $GLOBALS['TCA'][$table]['ctrl']['languageField'] . '<=0
-				OR
-				' . $GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField'] . ' = 0
-			)';
+            // Only restrict to the default language if no search request is in place
+            if ($this->searchString === '') {
+                $addWhere = (string)$queryBuilder->expr()->orX(
+                    $queryBuilder->expr()->lte($GLOBALS['TCA'][$table]['ctrl']['languageField'], 0),
+                    $queryBuilder->expr()->eq($GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField'], 0)
+                );
+            }
         }
         // Cleaning up:
         $this->fieldArray = array_unique(array_merge($this->fieldArray, $rowListArray));
@@ -203,7 +222,7 @@ class RecordListUtility extends DatabaseRecordList
         if ($thumbsCol) {
             $selectFields[] = $thumbsCol;
         }
-        if ($table == 'pages') {
+        if ($table === 'pages') {
             $selectFields[] = 'module';
             $selectFields[] = 'extendToSubpages';
             $selectFields[] = 'nav_hide';
@@ -215,7 +234,7 @@ class RecordListUtility extends DatabaseRecordList
         if (is_array($GLOBALS['TCA'][$table]['ctrl']['enablecolumns'])) {
             $selectFields = array_merge($selectFields, $GLOBALS['TCA'][$table]['ctrl']['enablecolumns']);
         }
-        foreach (array('type', 'typeicon_column', 'editlock') as $field) {
+        foreach (['type', 'typeicon_column', 'editlock'] as $field) {
             if ($GLOBALS['TCA'][$table]['ctrl'][$field]) {
                 $selectFields[] = $GLOBALS['TCA'][$table]['ctrl'][$field];
             }
@@ -239,12 +258,8 @@ class RecordListUtility extends DatabaseRecordList
         $selectFields = array_unique($selectFields);
         $fieldListFields = $this->makeFieldList($table, 1);
         if (empty($fieldListFields) && $GLOBALS['TYPO3_CONF_VARS']['BE']['debug']) {
-            $message = sprintf(
-                $lang->sL('LLL:EXT:lang/locallang_mod_web_list.xlf:missingTcaColumnsMessage', true),
-                $table,
-                $table
-            );
-            $messageTitle = $lang->sL('LLL:EXT:lang/locallang_mod_web_list.xlf:missingTcaColumnsMessageTitle', true);
+            $message = sprintf($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_mod_web_list.xlf:missingTcaColumnsMessage'), $table, $table);
+            $messageTitle = $lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_mod_web_list.xlf:missingTcaColumnsMessageTitle');
             /** @var FlashMessage $flashMessage */
             $flashMessage = GeneralUtility::makeInstance(
                 FlashMessage::class,
@@ -268,44 +283,56 @@ class RecordListUtility extends DatabaseRecordList
             foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['typo3/class.db_list_extra.inc']['getTable'] as $classData) {
                 $hookObject = GeneralUtility::getUserObj($classData);
                 if (!$hookObject instanceof RecordListGetTableHookInterface) {
-                    throw new \UnexpectedValueException(
-                        '$hookObject must implement interface ' . RecordListGetTableHookInterface::class,
-                        1195114460
-                    );
+                    throw new \UnexpectedValueException($classData . ' must implement interface ' . RecordListGetTableHookInterface::class, 1195114460);
                 }
                 $hookObject->getDBlistQuery($table, $id, $addWhere, $selFieldList, $this);
             }
         }
+        $selFieldList = GeneralUtility::trimExplode(',', $selFieldList, true);
 
-            // ingo.renner@dkd.de
-        if ($this->hideDisabledRecords) {
-            $addWhere .= ' AND '.$table.'.'
-                .$GLOBALS['TCA'][$table]['ctrl']['enablecolumns']['disabled']
-                .' = 0';
+        // ingo.renner@dkd.de
+        if ($this->hideDisabledRecords && isset($GLOBALS['TCA'][$table]['ctrl']['enablecolumns']['disabled'])) {
+            $addWhere = (string) $queryBuilder->expr()->andX(
+                $addWhere,
+                $queryBuilder->expr()->eq($GLOBALS['TCA'][$table]['ctrl']['enablecolumns']['disabled'], 0)
+            );
         }
 
-            //ingo.renner@dkd.de
-        if ($this->getBackendUserAuthentication()->user['admin'] == '0' && $table == 'be_users') {
-            $addWhere .= ' AND admin = 0';
-            $addWhere .= ' AND username NOT LIKE ("_cli%") ';
+        //ingo.renner@dkd.de
+        if (!$this->getBackendUserAuthentication()->isAdmin() && $table == 'be_users') {
+            $addWhere = (string) $queryBuilder->expr()->andX(
+                $addWhere,
+                $queryBuilder->expr()->eq('admin', 0),
+                $queryBuilder->expr()->notLike('username', '_cli%')
+            );
         }
 
         //dkd-kartolo
         //mod2, exclude fe_user which is also be_user
-        if ($table == 'fe_users') {
-            $addWhere .= ' AND username not in '.$this->excludeBE;
+        if ($table == 'fe_users' && !empty($this->excludeBE)) {
+            $addWhere = (string) $queryBuilder->expr()->andX(
+                $addWhere,
+                $queryBuilder->expr()->notIn('username', $this->excludeBE)
+            );
         }
 
         //dkd-kartolo
         //mod3, config dontShowPrefix
-        if ($table == 'be_groups' && $this->getBackendUserAuthentication()->user['admin']!= '1') {
+        if ($table == 'be_groups' && !$this->getBackendUserAuthentication()->isAdmin()) {
             $groupID = implode(',', TcBeuserUtility::showGroupID());
             if (!empty($groupID)) {
-                $addWhere .= ' AND uid in ('.$groupID.')';
+                $addWhere = (string) $queryBuilder->expr()->andX(
+                    $addWhere,
+                    $queryBuilder->expr()->in('uid', $groupID)
+                );
             } else {
-                $addWhere .= ' AND uid not in ('.TcBeuserUtility::getAllGroupsID().')';
+                $addWhere = (string) $queryBuilder->expr()->andX(
+                    $addWhere,
+                    $queryBuilder->expr()->notIn('uid', TcBeuserUtility::getAllGroupsID())
+                );
             }
         }
+        $additionalConstraints = empty($addWhere) ? [] : [QueryHelper::stripLogicalOperatorPrefix($addWhere)];
 
         // Create the SQL query for selecting the elements in the listing:
         // do not do paging when outputting as CSV
@@ -317,23 +344,23 @@ class RecordListUtility extends DatabaseRecordList
             $this->firstElementNumber = $this->firstElementNumber - 2;
             $this->iLimit = $this->iLimit + 2;
             // (API function from TYPO3\CMS\Recordlist\RecordList\AbstractDatabaseRecordList)
-            $queryParts = $this->makeQueryArray($table, $id, $addWhere, $selFieldList);
+            $queryBuilder = $this->getQueryBuilder($table, $id, $additionalConstraints);
             $this->firstElementNumber = $this->firstElementNumber + 2;
             $this->iLimit = $this->iLimit - 2;
         } else {
             // (API function from TYPO3\CMS\Recordlist\RecordList\AbstractDatabaseRecordList)
-            $queryParts = $this->makeQueryArray($table, $id, $addWhere, $selFieldList);
+            $queryBuilder = $this->getQueryBuilder($table, $id, $additionalConstraints);
         }
 
         // Finding the total amount of records on the page
         // (API function from TYPO3\CMS\Recordlist\RecordList\AbstractDatabaseRecordList)
-        $this->setTotalItems($queryParts);
+        $this->setTotalItems($table, $id, $additionalConstraints);
 
         // Init:
+        $queryResult = $queryBuilder->execute();
         $dbCount = 0;
         $out = '';
         $tableHeader = '';
-        $result = null;
         $listOnlyInSingleTableMode = $this->listOnlyInSingleTableMode && !$this->table;
         // If the count query returned any number of records, we perform the real query,
         // selecting records.
@@ -347,33 +374,25 @@ class RecordListUtility extends DatabaseRecordList
                     $this->showLimit = $this->totalItems;
                     $this->iLimit = $this->totalItems;
                 }
-                $result = $db->exec_SELECT_queryArray($queryParts);
-                $dbCount = $db->sql_num_rows($result);
+                $dbCount = $queryResult->rowCount();
             }
         }
         // If any records was selected, render the list:
         if ($dbCount) {
-            $tableTitle = $lang->sL($GLOBALS['TCA'][$table]['ctrl']['title'], true);
+            $tableTitle = htmlspecialchars($lang->sL($GLOBALS['TCA'][$table]['ctrl']['title']));
             if ($tableTitle === '') {
                 $tableTitle = $table;
             }
             // Header line is drawn
-            $theData = array();
+            $theData = [];
             if ($this->disableSingleTableView) {
                 $theData[$titleCol] = '<span class="c-table">' . BackendUtility::wrapInHelp($table, '', $tableTitle)
                     . '</span> (<span class="t3js-table-total-items">' . $this->totalItems . '</span>)';
             } else {
                 $icon = $this->table
-                    ? '<span title="' . $lang->getLL('contractView', true) . '">' .
-                        $this->iconFactory->getIcon('actions-view-table-collapse', Icon::SIZE_SMALL)->render() .
-                        '</span>'
-                    : '<span title="' . $lang->getLL('expandView', true) . '">' .
-                        $this->iconFactory->getIcon('actions-view-table-expand', Icon::SIZE_SMALL)->render() .
-                    '</span>';
-                $theData[$titleCol] = $this->linkWrapTable(
-                    $table,
-                    $tableTitle . ' (<span class="t3js-table-total-items">' . $this->totalItems . '</span>) ' . $icon
-                );
+                    ? '<span title="' . htmlspecialchars($lang->getLL('contractView')) . '">' . $this->iconFactory->getIcon('actions-view-table-collapse', Icon::SIZE_SMALL)->render() . '</span>'
+                    : '<span title="' . htmlspecialchars($lang->getLL('expandView')) . '">' . $this->iconFactory->getIcon('actions-view-table-expand', Icon::SIZE_SMALL)->render() . '</span>';
+                $theData[$titleCol] = $this->linkWrapTable($table, $tableTitle . ' (<span class="t3js-table-total-items">' . $this->totalItems . '</span>) ' . $icon);
             }
             if ($listOnlyInSingleTableMode) {
                 $tableHeader .= BackendUtility::wrapInHelp($table, '', $theData[$titleCol]);
@@ -381,24 +400,12 @@ class RecordListUtility extends DatabaseRecordList
                 // Render collapse button if in multi table mode
                 $collapseIcon = '';
                 if (!$this->table) {
-                    $href = htmlspecialchars(
-                        $this->listURL() . '&collapse[' . $table . ']=' . ($tableCollapsed ? '0' : '1')
-                    );
+                    $href = htmlspecialchars(($this->listURL() . '&collapse[' . $table . ']=' . ($tableCollapsed ? '0' : '1')));
                     $title = $tableCollapsed
-                        ? $lang->sL('LLL:EXT:lang/locallang_core.xlf:labels.expandTable', true)
-                        : $lang->sL('LLL:EXT:lang/locallang_core.xlf:labels.collapseTable', true);
-                    $icon = '<span class="collapseIcon">' .
-                        $this->iconFactory->getIcon(
-                            ($tableCollapsed ? 'actions-view-list-expand' : 'actions-view-list-collapse'),
-                            Icon::SIZE_SMALL
-                        )->render() .
-                        '</span>';
-                    $collapseIcon = '<a href="' . $href . '"' .
-                        ' title="' . $title .  '"' .
-                        ' class="pull-right t3js-toggle-recordlist"' .
-                        ' data-table="' . htmlspecialchars($table) . '"' .
-                        ' data-toggle="collapse" data-target="#recordlist-' . htmlspecialchars($table) . '">'
-                        . $icon . '</a>';
+                        ? htmlspecialchars($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels.expandTable'))
+                        : htmlspecialchars($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels.collapseTable'));
+                    $icon = '<span class="collapseIcon">' . $this->iconFactory->getIcon(($tableCollapsed ? 'actions-view-list-expand' : 'actions-view-list-collapse'), Icon::SIZE_SMALL)->render() . '</span>';
+                    $collapseIcon = '<a href="' . $href . '" title="' . $title . '" class="pull-right t3js-toggle-recordlist" data-table="' . htmlspecialchars($table) . '" data-toggle="collapse" data-target="#recordlist-' . htmlspecialchars($table) . '">' . $icon . '</a>';
                 }
                 $tableHeader .= $theData[$titleCol] . $collapseIcon;
             }
@@ -406,21 +413,21 @@ class RecordListUtility extends DatabaseRecordList
             $rowOutput = '';
             if (!$listOnlyInSingleTableMode || $this->table) {
                 // Fixing an order table for sortby tables
-                $this->currentTable = array();
-                $currentIdList = array();
+                $this->currentTable = [];
+                $currentIdList = [];
                 $doSort = $GLOBALS['TCA'][$table]['ctrl']['sortby'] && !$this->sortField;
                 $prevUid = 0;
                 $prevPrevUid = 0;
                 // Get first two rows and initialize prevPrevUid and prevUid if on page > 1
                 if ($this->firstElementNumber > 2 && $this->iLimit > 0) {
-                    $row = $db->sql_fetch_assoc($result);
+                    $row = $queryResult->fetch();
                     $prevPrevUid = -((int)$row['uid']);
-                    $row = $db->sql_fetch_assoc($result);
+                    $row = $queryResult->fetch();
                     $prevUid = $row['uid'];
                 }
-                $accRows = array();
+                $accRows = [];
                 // Accumulate rows here
-                while ($row = $db->sql_fetch_assoc($result)) {
+                while ($row = $queryResult->fetch()) {
                     if (!$this->isRowListingConditionFulfilled($table, $row)) {
                         continue;
                     }
@@ -440,15 +447,14 @@ class RecordListUtility extends DatabaseRecordList
                         }
                     }
                 }
-                $db->sql_free_result($result);
                 $this->totalRowCount = count($accRows);
                 // CSV initiated
                 if ($this->csvOutput) {
                     $this->initCSV();
                 }
                 // Render items:
-                $this->CBnames = array();
-                $this->duplicateStack = array();
+                $this->CBnames = [];
+                $this->duplicateStack = [];
                 $this->eCounter = $this->firstElementNumber;
                 $cc = 0;
                 foreach ($accRows as $row) {
@@ -457,27 +463,47 @@ class RecordListUtility extends DatabaseRecordList
                         $cc++;
                         $this->translations = false;
                         $rowOutput .= $this->renderListRow($table, $row, $cc, $titleCol, $thumbsCol);
-                        // If localization view is enabled it means that the selected records are
-                        // either default or All language and here we will not select translations
+                        // If localization view is enabled and no search happened it means that the selected
+                        // records are either default or All language and here we will not select translations
                         // which point to the main record:
-                        if ($this->localizationView && $l10nEnabled) {
+                        if ($this->localizationView && $l10nEnabled && $this->searchString === '') {
                             // For each available translation, render the record:
                             if (is_array($this->translations)) {
                                 foreach ($this->translations as $lRow) {
                                     // $lRow isn't always what we want - if record was moved we've to work with the
                                     // placeholder records otherwise the list is messed up a bit
                                     if ($row['_MOVE_PLH_uid'] && $row['_MOVE_PLH_pid']) {
-                                        $where = 't3ver_move_id="' . (int)$lRow['uid'] . '"' .
-                                        ' AND pid="' . $row['_MOVE_PLH_pid'] . '"' .
-                                        ' AND t3ver_wsid=' . $row['t3ver_wsid'] . BackendUtility::deleteClause($table);
-                                        $tmpRow = BackendUtility::getRecordRaw($table, $where, $selFieldList);
+                                        $queryBuilder = $this->getQueryBuilderForTable($table);
+                                        $queryBuilder->getRestrictions()
+                                            ->removeAll()
+                                            ->add(GeneralUtility::makeInstance(DeletedRestriction::class));
+                                        $predicates = [
+                                            $queryBuilder->expr()->eq(
+                                                't3ver_move_id',
+                                                $queryBuilder->createNamedParameter((int)$lRow['uid'], \PDO::PARAM_INT)
+                                            ),
+                                            $queryBuilder->expr()->eq(
+                                                'pid',
+                                                $queryBuilder->createNamedParameter((int)$row['_MOVE_PLH_pid'], \PDO::PARAM_INT)
+                                            ),
+                                            $queryBuilder->expr()->eq(
+                                                't3ver_wsid',
+                                                $queryBuilder->createNamedParameter((int)$row['t3ver_wsid'], \PDO::PARAM_INT)
+                                            ),
+                                        ];
+
+                                        $tmpRow = $queryBuilder
+                                            ->select(...$selFieldList)
+                                            ->from($table)
+                                            ->andWhere(...$predicates)
+                                            ->execute()
+                                            ->fetch();
+
                                         $lRow = is_array($tmpRow) ? $tmpRow : $lRow;
                                     }
                                     // In offline workspace, look for alternative record:
                                     BackendUtility::workspaceOL($table, $lRow, $backendUser->workspace, true);
-                                    if (is_array($lRow) &&
-                                        $backendUser->checkLanguageAccess($lRow[$GLOBALS['TCA'][$table]['ctrl']['languageField']])
-                                    ) {
+                                    if (is_array($lRow) && $backendUser->checkLanguageAccess($lRow[$GLOBALS['TCA'][$table]['ctrl']['languageField']])) {
                                         $currentIdList[] = $lRow['uid'];
                                         $rowOutput .= $this->renderListRow($table, $lRow, $cc, $titleCol, $thumbsCol, 18);
                                     }
@@ -491,9 +517,7 @@ class RecordListUtility extends DatabaseRecordList
                 // Record navigation is added to the beginning and end of the table if in single
                 // table mode
                 if ($this->table) {
-                    $rowOutput = $this->renderListNavigation('top') .
-                        $rowOutput .
-                        $this->renderListNavigation('bottom');
+                    $rowOutput = $this->renderListNavigation('top') . $rowOutput . $this->renderListNavigation('bottom');
                 } else {
                     // Show that there are more records than shown
                     if ($this->totalItems > $this->itemsLimitPerTable) {
@@ -501,9 +525,9 @@ class RecordListUtility extends DatabaseRecordList
                         $hasMore = $this->totalItems > $this->itemsLimitSingleTable;
                         $colspan = $this->showIcon ? count($this->fieldArray) + 1 : count($this->fieldArray);
                         $rowOutput .= '<tr><td colspan="' . $colspan . '">
-                            <a href="' . htmlspecialchars(($this->listURL() . '&table=' . rawurlencode($table))) . '" class="btn btn-default">' .
-                            '<span class="t3-icon fa fa-chevron-down"></span> <i>[1 - ' . $countOnFirstPage . ($hasMore ? '+' : '') . ']</i></a>
-                            </td></tr>';
+								<a href="' . htmlspecialchars(($this->listURL() . '&table=' . rawurlencode($table))) . '" class="btn btn-default">'
+                            . '<span class="t3-icon fa fa-chevron-down"></span> <i>[1 - ' . $countOnFirstPage . ($hasMore ? '+' : '') . ']</i></a>
+								</td></tr>';
                     }
                 }
                 // The header row for the table is now created:
@@ -550,7 +574,8 @@ class RecordListUtility extends DatabaseRecordList
      * Rendering the header row for a table
      *
      * @param string $table Table name
-     * @param array $currentIdList Array of the currectly displayed uids of the table
+     * @param int[] $currentIdList Array of the currently displayed uids of the table
+     * @throws \UnexpectedValueException
      * @return string Header table row
      * @access private
      * @see getTable()
@@ -559,37 +584,36 @@ class RecordListUtility extends DatabaseRecordList
     {
         $lang = $this->getLanguageService();
         // Init:
-        $theData = array();
+        $theData = [];
         $icon = '';
         // Traverse the fields:
         foreach ($this->fieldArray as $fCol) {
             // Calculate users permissions to edit records in the table:
-            $permsEdit = $this->calcPerms & ($table == 'pages' ? 2 : 16) && $this->overlayEditLockPermissions($table);
+            $permsEdit = $this->calcPerms & ($table === 'pages' ? 2 : 16) && $this->overlayEditLockPermissions($table);
             switch ((string)$fCol) {
                 case '_PATH_':
                     // Path
-                    $theData[$fCol] = '<i>[' . $lang->sL('LLL:EXT:lang/locallang_core.xlf:labels._PATH_', true) . ']</i>';
+                    $theData[$fCol] = '<i>[' . htmlspecialchars($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels._PATH_')) . ']</i>';
                     break;
                 case '_REF_':
                     // References
-                    //$theData[$fCol] = '<i>[' . $lang->sL('LLL:EXT:lang/locallang_mod_file_list.xlf:c__REF_', true) . ']</i>';
+                    $theData[$fCol] = '<i>[' . htmlspecialchars($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_mod_file_list.xlf:c__REF_')) . ']</i>';
                     break;
                 case '_LOCALIZATION_':
                     // Path
-                    $theData[$fCol] = '<i>[' . $lang->sL('LLL:EXT:lang/locallang_core.xlf:labels._LOCALIZATION_', true) . ']</i>';
+                    $theData[$fCol] = '<i>[' . htmlspecialchars($lang->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels._LOCALIZATION_')) . ']</i>';
                     break;
                 case '_LOCALIZATION_b':
                     // Path
-                    $theData[$fCol] = $lang->getLL('Localize', true);
+                    $theData[$fCol] = htmlspecialchars($lang->getLL('Localize'));
                     break;
                 case '_CLIPBOARD_':
                     if (!$this->getModule()->MOD_SETTINGS['clipBoard']) {
                         break;
                     }
                     // Clipboard:
-                    $cells = array();
-                    // If there are elements on the clipboard for this table,
-                    // and the parent page is not locked by editlock
+                    $cells = [];
+                    // If there are elements on the clipboard for this table, and the parent page is not locked by editlock
                     // then display the "paste into" icon:
                     $elFromTable = $this->clipObj->elFromTable($table);
                     if (!empty($elFromTable) && $this->overlayEditLockPermissions($table)) {
@@ -597,30 +621,25 @@ class RecordListUtility extends DatabaseRecordList
                         $confirmMessage = $this->clipObj->confirmMsgText('pages', $this->pageRow, 'into', $elFromTable);
                         $cells['pasteAfter'] = '<a class="btn btn-default t3js-modal-trigger"'
                             . ' href="' . $href . '"'
-                            . ' title="' . $lang->getLL('clip_paste', true) . '"'
-                            . ' data-title="' . $lang->getLL('clip_paste', true) . '"'
+                            . ' title="' . htmlspecialchars($lang->getLL('clip_paste')) . '"'
+                            . ' data-title="' . htmlspecialchars($lang->getLL('clip_paste')) . '"'
                             . ' data-content="' . htmlspecialchars($confirmMessage) . '"'
                             . ' data-severity="warning">'
-                            . $this->iconFactory->getIcon('actions-document-paste-after', Icon::SIZE_SMALL)->render()
+                            . $this->iconFactory->getIcon('actions-document-paste-into', Icon::SIZE_SMALL)->render()
                             . '</a>';
                     }
                     // If the numeric clipboard pads are enabled, display the control icons for that:
-                    if ($this->clipObj->current != 'normal') {
+                    if ($this->clipObj->current !== 'normal') {
                         // The "select" link:
                         $spriteIcon = $this->iconFactory->getIcon('actions-edit-copy', Icon::SIZE_SMALL)->render();
                         $cells['copyMarked'] = $this->linkClipboardHeaderIcon($spriteIcon, $table, 'setCB', '', $lang->getLL('clip_selectMarked'));
-
                         // The "edit marked" link:
-                        $editIdList = implode(',', $currentIdList);
-                        $editIdList = '\'+editList(' . GeneralUtility::quoteJSvalue($table) . ',' . GeneralUtility::quoteJSvalue($editIdList) . ')+\'';
-                        $params = 'edit[' . $table . '][' . $editIdList . ']=edit';
-                        $onClick = BackendUtility::editOnClick('', '', -1);
-                        $onClickArray = explode('?', $onClick, 2);
-                        $lastElement = array_pop($onClickArray);
-                        array_push($onClickArray, $params . '&' . $lastElement);
-                        $onClick = implode('?', $onClickArray);
-                        $cells['edit'] = '<a class="btn btn-default" href="#" onclick="' . htmlspecialchars($onClick) . '" title="'
-                            . $lang->getLL('clip_editMarked', true) . '">'
+                        $editUri = BackendUtility::getModuleUrl('record_edit')
+                            . '&edit[' . $table . '][{entityIdentifiers:editList}]=edit'
+                            . '&returnUrl={T3_THIS_LOCATION}';
+                        $cells['edit'] = '<a class="btn btn-default t3js-record-edit-multiple" href="#"'
+                            . ' data-uri="' . htmlspecialchars($editUri) . '"'
+                            . ' title="' . htmlspecialchars($lang->getLL('clip_editMarked')) . '">'
                             . $this->iconFactory->getIcon('actions-document-open', Icon::SIZE_SMALL)->render() . '</a>';
                         // The "Delete marked" link:
                         $cells['delete'] = $this->linkClipboardHeaderIcon(
@@ -633,7 +652,7 @@ class RecordListUtility extends DatabaseRecordList
                         // The "Select all" link:
                         $onClick = htmlspecialchars(('checkOffCB(' . GeneralUtility::quoteJSvalue(implode(',', $this->CBnames)) . ', this); return false;'));
                         $cells['markAll'] = '<a class="btn btn-default" rel="" href="#" onclick="' . $onClick . '" title="'
-                            . $lang->getLL('clip_markRecords', true) . '">'
+                            . htmlspecialchars($lang->getLL('clip_markRecords')) . '">'
                             . $this->iconFactory->getIcon('actions-document-select', Icon::SIZE_SMALL)->render() . '</a>';
                     } else {
                         $cells['empty'] = '';
@@ -644,17 +663,22 @@ class RecordListUtility extends DatabaseRecordList
                      *        This hook allows to modify the icons responsible for the clipboard functions
                      *        (shown above the clipboard checkboxes when a clipboard other than "Normal" is selected),
                      *        or other "Action" functions which perform operations on the listed records.
-                     */
+                    */
                     if (is_array($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['typo3/class.db_list_extra.inc']['actions'])) {
                         foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['typo3/class.db_list_extra.inc']['actions'] as $classData) {
                             $hookObject = GeneralUtility::getUserObj($classData);
                             if (!$hookObject instanceof RecordListHookInterface) {
-                                throw new \UnexpectedValueException('$hookObject must implement interface ' . RecordListHookInterface::class, 1195567850);
+                                throw new \UnexpectedValueException($classData . ' must implement interface ' . RecordListHookInterface::class, 1195567850);
                             }
                             $cells = $hookObject->renderListHeaderActions($table, $currentIdList, $cells, $this);
                         }
                     }
-                    $theData[$fCol] = '<div class="btn-group" role="group">' . implode('', $cells) . '</div>';
+                    $theData[$fCol] = '';
+                    if (isset($cells['edit']) && isset($cells['delete'])) {
+                        $theData[$fCol] .= '<div class="btn-group" role="group">' . $cells['edit'] . $cells['delete'] . '</div>';
+                        unset($cells['edit'], $cells['delete']);
+                    }
+                    $theData[$fCol] .= '<div class="btn-group" role="group">' . implode('', $cells) . '</div>';
                     break;
                 case '_CONTROL_':
                     // dkd-kartolo: dummy header
@@ -668,7 +692,7 @@ class RecordListUtility extends DatabaseRecordList
                     // at the end
                     $sortLabel = BackendUtility::getItemLabel($table, $fCol);
                     if ($sortLabel !== null) {
-                        $sortLabel = $lang->sL($sortLabel, true);
+                        $sortLabel = htmlspecialchars($lang->sL($sortLabel));
                         $sortLabel = rtrim(trim($sortLabel), ':');
                     } else {
                         // No TCA field, only output the $fCol variable with square brackets []
@@ -680,7 +704,7 @@ class RecordListUtility extends DatabaseRecordList
                         // If the numeric clipboard pads are selected, show duplicate sorting link:
                         if ($this->clipNumPane()) {
                             $theData[$fCol] .= '<a class="btn btn-default" href="' . htmlspecialchars($this->listURL('', '-1') . '&duplicateField=' . $fCol)
-                                . '" title="' . $lang->getLL('clip_duplicates', true) . '">'
+                                . '" title="' . htmlspecialchars($lang->getLL('clip_duplicates')) . '">'
                                 . $this->iconFactory->getIcon('actions-document-duplicates-select', Icon::SIZE_SMALL)->render() . '</a>';
                         }
                         if (strlen($theData[$fCol]) > 0) {
@@ -690,7 +714,6 @@ class RecordListUtility extends DatabaseRecordList
                     $theData[$fCol] .= $this->addSortLink($sortLabel, $fCol, $table);
             }
         }
-
         /**
          * @hook renderListHeader: Allows to change the contents of columns/cells of the Web>List table headers
          * @usage Above each listed table in Web>List a header row is shown.
@@ -701,7 +724,7 @@ class RecordListUtility extends DatabaseRecordList
             foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['typo3/class.db_list_extra.inc']['actions'] as $classData) {
                 $hookObject = GeneralUtility::getUserObj($classData);
                 if (!$hookObject instanceof RecordListHookInterface) {
-                    throw new \UnexpectedValueException('$hookObject must implement interface ' . RecordListHookInterface::class, 1195567855);
+                    throw new \UnexpectedValueException($classData . ' must implement interface ' . RecordListHookInterface::class, 1195567855);
                 }
                 $theData = $hookObject->renderListHeader($table, $currentIdList, $theData, $this);
             }
@@ -714,13 +737,13 @@ class RecordListUtility extends DatabaseRecordList
     /**
      * Rendering a single row for the list
      *
-     * @param    string $table : Table name
-     * @param    array $row : Current record
-     * @param    integer $cc : Counter, counting for each time an element is rendered (used for alternating colors)
-     * @param    string $titleCol : Table field (column) where header value is found
-     * @param    string $thumbsCol : Table field (column) where (possible) thumbnails can be found
-     * @param    int $indent : indent from left.
-     * @return    string        Table row for the element
+     * @param string $table Table name
+     * @param mixed[] $row Current record
+     * @param int $cc Counter, counting for each time an element is rendered (used for alternating colors)
+     * @param string $titleCol Table field (column) where header value is found
+     * @param string $thumbsCol Table field (column) where (possible) thumbnails can be found
+     * @param int $indent Indent from left.
+     * @return string Table row for the element
      * @access private
      * @see getTable()
      */
@@ -736,19 +759,24 @@ class RecordListUtility extends DatabaseRecordList
             $id_orig = $this->id;
             $this->id = $row['pid'];
         }
+
+        $tagAttributes = [
+            'class' => ['t3js-entity'],
+            'data-table' => $table,
+            'title' => 'id=' . $row['uid'],
+        ];
+
         // Add special classes for first and last row
-        $rowSpecial = '';
         if ($cc == 1 && $indent == 0) {
-            $rowSpecial .= ' firstcol';
+            $tagAttributes['class'][] = 'firstcol';
         }
         if ($cc == $this->totalRowCount || $cc == $this->iLimit) {
-            $rowSpecial .= ' lastcol';
+            $tagAttributes['class'][] = 'lastcol';
         }
-
-        $row_bgColor = ' class="' . $rowSpecial . '"';
-
         // Overriding with versions background color if any:
-        $row_bgColor = $row['_CSSCLASS'] ? ' class="' . $row['_CSSCLASS'] . '"' : $row_bgColor;
+        if (!empty($row['_CSSCLASS'])) {
+            $tagAttributes['class'] = [$row['_CSSCLASS']];
+        }
         // Incr. counter.
         $this->counter++;
         // The icon with link
@@ -759,9 +787,8 @@ class RecordListUtility extends DatabaseRecordList
             . '</span>';
         $theIcon = $this->clickMenuEnabled ? BackendUtility::wrapClickMenuOnIcon($iconImg, $table, $row['uid']) : $iconImg;
         // Preparing and getting the data-array
-        $theData = array();
+        $theData = [];
         $localizationMarkerClass = '';
-
         foreach ($this->fieldArray as $fCol) {
             if ($fCol == $titleCol) {
                 $recTitle = BackendUtility::getRecordTitle($table, $row, false, true);
@@ -769,10 +796,8 @@ class RecordListUtility extends DatabaseRecordList
                 // If the record is edit-locked	by another user, we will show a little warning sign:
                 $lockInfo = BackendUtility::isRecordLocked($table, $row['uid']);
                 if ($lockInfo) {
-                    $warning = '<a href="#" onclick="alert('
-                        . GeneralUtility::quoteJSvalue($lockInfo['msg']) . '); return false;" title="'
-                        . htmlspecialchars($lockInfo['msg']) . '">'
-                        . $this->iconFactory->getIcon('status-warning-in-use', Icon::SIZE_SMALL)->render() . '</a>';
+                    $warning = '<span data-toggle="tooltip" data-placement="right" data-title="' . htmlspecialchars($lockInfo['msg']) . '">'
+                        . $this->iconFactory->getIcon('status-warning-in-use', Icon::SIZE_SMALL)->render() . '</span>';
                 }
                 $theData[$fCol] = $theData['__label'] = $warning . $this->linkWrapItems($table, $row['uid'], $recTitle, $row);
                 // Render thumbnails, if:
@@ -795,7 +820,9 @@ class RecordListUtility extends DatabaseRecordList
                     trim($row[$thumbsCol]) &&
                     preg_match('/(^|(.*(;|,)?))' . $thumbsCol . '(((;|,).*)|$)/', $visibleColumns) === 1
                 ) {
-                    $theData[$fCol] .= '<br />' . $this->thumbCode($row, $table, $thumbsCol);
+                    $thumbCode = '<br />' . $this->thumbCode($row, $table, $thumbsCol);
+                    $theData[$fCol] .= $thumbCode;
+                    $theData['__label'] .= $thumbCode;
                 }
                 if (isset($GLOBALS['TCA'][$table]['ctrl']['languageField'])
                     && $row[$GLOBALS['TCA'][$table]['ctrl']['languageField']] != 0
@@ -804,21 +831,21 @@ class RecordListUtility extends DatabaseRecordList
                     // It's a translated record with a language parent
                     $localizationMarkerClass = ' localization';
                 }
-            } elseif ($fCol == 'pid') {
+            } elseif ($fCol === 'pid') {
                 $theData[$fCol] = $row[$fCol];
-            } elseif ($fCol == '_PATH_') {
+            } elseif ($fCol === '_PATH_') {
                 $theData[$fCol] = $this->recPath($row['pid']);
-            } elseif ($fCol == '_REF_') {
+            } elseif ($fCol === '_REF_') {
                 $theData[$fCol] = $this->createReferenceHtml($table, $row['uid']);
-            } elseif ($fCol == '_CONTROL_') {
+            } elseif ($fCol === '_CONTROL_') {
                 $theData[$fCol] = $this->makeControl($table, $row);
-            } elseif ($fCol == '_CLIPBOARD_') {
+            } elseif ($fCol === '_CLIPBOARD_') {
                 $theData[$fCol] = $this->makeClip($table, $row);
-            } elseif ($fCol == '_LOCALIZATION_') {
+            } elseif ($fCol === '_LOCALIZATION_') {
                 list($lC1, $lC2) = $this->makeLocalizationPanel($table, $row);
                 $theData[$fCol] = $lC1;
                 $theData[$fCol . 'b'] = '<div class="btn-group">' . $lC2 . '</div>';
-            } elseif ($fCol == '_LOCALIZATION_b') {
+            } elseif ($fCol === '_LOCALIZATION_b') {
                 // deliberately empty
             } elseif ($this->userMainGroupOnly && $table == 'be_users' && $fCol == 'usergroup') {
                 // ingo.renner@dkd.de
@@ -841,7 +868,8 @@ class RecordListUtility extends DatabaseRecordList
             $this->addToCSV($row);
         }
         // Add classes to table cells
-//        $this->addElement_tdCssClass[$titleCol] = 'col-title' . $localizationMarkerClass;
+//        $this->addElement_tdCssClass[$titleCol] = 'col-title col-responsive' . $localizationMarkerClass;
+        $this->addElement_tdCssClass['__label'] = $this->addElement_tdCssClass[$titleCol];
         $this->addElement_tdCssClass['_CONTROL_'] = 'col-control';
         if ($this->getModule()->MOD_SETTINGS['clipBoard']) {
             $this->addElement_tdCssClass['_CLIPBOARD_'] = 'col-clipboard';
@@ -853,27 +881,45 @@ class RecordListUtility extends DatabaseRecordList
         $theData['uid'] = $row['uid'];
         if (isset($GLOBALS['TCA'][$table]['ctrl']['languageField'])
             && isset($GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField'])
-            && !isset($GLOBALS['TCA'][$table]['ctrl']['transOrigPointerTable'])
+            && $table !== 'pages_language_overlay'
         ) {
-            $theData['parent'] = $row[$GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']];
+            $theData['_l10nparent_'] = $row[$GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']];
         }
-        $rowOutput .= $this->addElement(1, $theIcon, $theData, $row_bgColor);
+
+        $tagAttributes = array_map(
+            function ($attributeValue) {
+                if (is_array($attributeValue)) {
+                    return implode(' ', $attributeValue);
+                }
+                return $attributeValue;
+            },
+            $tagAttributes
+        );
+
+        $rowOutput .= $this->addElement(1, $theIcon, $theData, GeneralUtility::implodeAttributes($tagAttributes, true));
         // Finally, return table row element:
         return $rowOutput;
     }
 
+    /**
+     * Returns the user's main group
+     *
+     * @param string $allGroups
+     * @return string
+     */
     public function getUserMainGroup($allGroups)
     {
         $allGroups = explode(',', $allGroups);
         if (!empty($allGroups[0])) {
-            $mainGroup = $this->getDatabaseConnection()->exec_SELECTgetRows(
-                'title',
-                'be_groups',
-                'uid = '.$allGroups[0]
-            );
+            $queryBuilder = $this->getQueryBuilderForTable('be_groups');
+            $mainGroup = $queryBuilder->select('*')
+                ->from('be_groups')
+                ->where($queryBuilder->expr()->eq('uid', $allGroups[0]))
+                ->execute()
+                ->fetch();
         }
 
-        return $mainGroup[0]['title'];
+        return (string) ($mainGroup ? $mainGroup['title'] : '');
     }
 
     /**
@@ -903,7 +949,8 @@ class RecordListUtility extends DatabaseRecordList
      * Creates the control panel for a single record in the listing.
      *
      * @param string $table The table
-     * @param array $row The record for which to make the control panel.
+     * @param mixed[] $row The record for which to make the control panel.
+     * @throws \UnexpectedValueException
      * @return string HTML table with the control panel (unless disabled)
      */
     public function makeControl($table, $row)
@@ -913,14 +960,13 @@ class RecordListUtility extends DatabaseRecordList
         }
 
         $rowUid = $row['uid'];
-
-        $cells = array(
-            'primary' => array(),
-            'secondary' => array()
-        );
-
+        $cells = [
+            'primary' => [],
+            'secondary' => []
+        ];
         // If the listed table is 'pages' we have to request the permission settings for each page:
-        if ($table == 'pages') {
+        $localCalcPerms = 0;
+        if ($table === 'pages') {
             $localCalcPerms = $this->getBackendUserAuthentication()->calcPerms(BackendUtility::getRecord('pages', $row['uid']));
         }
 
@@ -933,16 +979,18 @@ class RecordListUtility extends DatabaseRecordList
             if ($permsEdit) {
                 $params = '&edit[' . $table . '][' . $row['uid'] . ']=edit';
                 $iconIdentifier = 'actions-open';
+                if ($table === 'pages') {
+                    $iconIdentifier = 'actions-page-open';
+                }
                 $overlayIdentifier = !$this->isEditable($table) ? 'overlay-readonly' : null;
-                $editAction = '<a class="btn btn-default" href="#" onclick="' . htmlspecialchars(self::editOnClick($params, '', -1))
-                    . '" title="' . $this->getLanguageService()->getLL('edit', true) . '">' .
-                    $this->iconFactory->getIcon($iconIdentifier, Icon::SIZE_SMALL, $overlayIdentifier)->render() .
-                    '</a>';
+                $editAction = '<a class="btn btn-default" href="#" onclick="' . htmlspecialchars(BackendUtility::editOnClick($params, '', -1))
+                    . '" title="' . htmlspecialchars($this->getLanguageService()->getLL('edit')) . '">' . $this->iconFactory->getIcon($iconIdentifier, Icon::SIZE_SMALL, $overlayIdentifier)->render() . '</a>';
             } else {
                 $editAction = $this->spaceIcon;
             }
             $this->addActionToCellGroup($cells, $editAction, 'edit');
         }
+
 
             //dkd-kartolo
             //show magnifier (mod4)
@@ -970,13 +1018,14 @@ class RecordListUtility extends DatabaseRecordList
 
         // "Info": (All records)
         $onClick = 'top.launchView(' . GeneralUtility::quoteJSvalue($table) . ', ' . (int)$row['uid'] . '); return false;';
-        $viewBigAction = '<a class="btn btn-default" href="#" onclick="' . htmlspecialchars($onClick) . '" title="' . $this->getLanguageService()->getLL('showInfo', true) . '">'
+        $viewBigAction = '<a class="btn btn-default" href="#" onclick="' . htmlspecialchars($onClick) . '" title="' . htmlspecialchars($this->getLanguageService()->getLL('showInfo')) . '">'
             . $this->iconFactory->getIcon('actions-document-info', Icon::SIZE_SMALL)->render() . '</a>';
         $this->addActionToCellGroup($cells, $viewBigAction, 'viewBig');
 
-        // "Hide/Unhide" links:
+            // "Hide/Unhide" links:
         if (!$this->disableControls['hide']) {
             $hiddenField = $GLOBALS['TCA'][$table]['ctrl']['enablecolumns']['disabled'];
+
             if ($permsEdit && $hiddenField && $GLOBALS['TCA'][$table]['columns'][$hiddenField]
                 && (!$GLOBALS['TCA'][$table]['columns'][$hiddenField]['exclude']
                     || $this->getBackendUserAuthentication()->check('non_exclude_fields', $table . ':' . $hiddenField))
@@ -984,30 +1033,29 @@ class RecordListUtility extends DatabaseRecordList
                 if ($this->isRecordCurrentBackendUser($table, $row)) {
                     $hideAction = $this->spaceIcon;
                 } else {
-                    $hideTitle = $this->getLanguageService()->getLL('hide' . ($table == 'pages' ? 'Page' : ''), true);
-                    $unhideTitle = $this->getLanguageService()->getLL('unHide' . ($table == 'pages' ? 'Page' : ''), true);
-
+                    $hideTitle = htmlspecialchars($this->getLanguageService()->getLL('hide' . ($table === 'pages' ? 'Page' : '')));
+                    $unhideTitle = htmlspecialchars($this->getLanguageService()->getLL('unHide' . ($table === 'pages' ? 'Page' : '')));
                     if ($row[$hiddenField]) {
-                        $params = '&data[' . $table . '][' . $row['uid'] . '][' . $hiddenField . ']=0&SET[function]=action';
-                        $hideAction = '<a href="#" class="btn btn-default" ' .
-                            ' title="' . $unhideTitle . '" ' .
-                            'onclick="' . htmlspecialchars('return jumpToUrl(\'' . self::actionOnClick($params) . '\');') . '">' .
-                            $this->iconFactory->getIcon('actions-edit-unhide', Icon::SIZE_SMALL)->render() .
-                            '</a>';
+                        $params = 'data[' . $table . '][' . $rowUid . '][' . $hiddenField . ']=0';
+                        $hideAction = '<a class="btn btn-default t3js-record-hide" data-state="hidden" href="#"'
+                                      . ' data-params="' . htmlspecialchars($params) . '"'
+                                      . ' title="' . $unhideTitle . '"'
+                                      . ' data-toggle-title="' . $hideTitle . '">'
+                                      . $this->iconFactory->getIcon('actions-edit-unhide', Icon::SIZE_SMALL)->render() . '</a>';
                     } else {
-                        $params = '&data[' . $table . '][' . $row['uid'] . '][' . $hiddenField . ']=1&SET[function]=action';
-                        $hideAction = '<a href="#" class="btn btn-default" ' .
-                            'title="' . $hideTitle . '" ' .
-                            'onclick=" ' . htmlspecialchars('return jumpToUrl(\'' . self::actionOnClick($params) . '\');') . '">' .
-                            $this->iconFactory->getIcon('actions-edit-hide', Icon::SIZE_SMALL)->render() .
-                            '</a>';
+                        $params = 'data[' . $table . '][' . $rowUid . '][' . $hiddenField . ']=1';
+                        $hideAction = '<a class="btn btn-default t3js-record-hide" data-state="visible" href="#"'
+                                      . ' data-params="' . htmlspecialchars($params) . '"'
+                                      . ' title="' . $hideTitle . '"'
+                                      . ' data-toggle-title="' . $unhideTitle . '">'
+                                      . $this->iconFactory->getIcon('actions-edit-hide', Icon::SIZE_SMALL)->render() . '</a>';
                     }
                 }
                 $this->addActionToCellGroup($cells, $hideAction, 'hide');
             }
         }
 
-        // "Delete" link:
+            // "Delete" link:
         if (!$this->disableControls['delete']) {
             if ($permsEdit && ($table === 'pages' && $localCalcPerms & Permission::PAGE_DELETE || $table !== 'pages' && $this->calcPerms & Permission::CONTENT_EDIT)) {
                 // Check if the record version is in "deleted" state, because that will switch the action to "restore"
@@ -1019,35 +1067,29 @@ class RecordListUtility extends DatabaseRecordList
                     $refCountMsg = BackendUtility::referenceCount(
                         $table,
                         $row['uid'],
-                        ' ' .
-                        $this->getLanguageService()->sL('LLL:EXT:lang/locallang_core.xlf:labels.referencesToRecord'),
-                        $this->getReferenceCount(
+                        ' ' . $this->getLanguageService()->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels.referencesToRecord'),
+                        $this->getReferenceCount($table, $row['uid'])
+                    ) . BackendUtility::translationCount(
                             $table,
-                            $row['uid']
-                        )
-                    ) .
-                    BackendUtility::translationCount(
-                        $table,
-                        $row['uid'],
-                        ' ' .
-                        $this->getLanguageService()->sL('LLL:EXT:lang/locallang_core.xlf:labels.translationsOfRecord')
+                            $row['uid'],
+                        ' ' . $this->getLanguageService()->sL('LLL:EXT:lang/Resources/Private/Language/locallang_core.xlf:labels.translationsOfRecord')
                     );
                 }
 
                 if ($this->isRecordCurrentBackendUser($table, $row)) {
                     $deleteAction = $this->spaceIcon;
                 } else {
-                    $titleOrig = BackendUtility::getRecordTitle($table, $row, false, true);
-                    $title = GeneralUtility::slashJS(GeneralUtility::fixed_lgd_cs($titleOrig, $this->fixedL), true);
+                    $title = BackendUtility::getRecordTitle($table, $row);
                     $warningText = $this->getLanguageService()->getLL($actionName . 'Warning') . ' "' . $title . '" ' . '[' . $table . ':' . $row['uid'] . ']' . $refCountMsg;
-                    $linkTitle = $this->getLanguageService()->getLL($actionName, true);
 
-                    $params = '&cmd[' . $table . '][' . $row['uid'] . '][delete]=1&SET[function]=action';
-                    $deleteAction = '<a href="#" class="btn btn-default"' .
-                        ' title="' . $linkTitle . '"' .
-                        ' onclick="' . htmlspecialchars('if (confirm(' . GeneralUtility::quoteJSvalue($warningText) . ')) {jumpToUrl(\'' . $this->actionOnClick($params) . '\');}') . '">' .
-                        $this->iconFactory->getIcon('actions-edit-' . $actionName, Icon::SIZE_SMALL)->render() .
-                        '</a>';
+                    $params = 'cmd[' . $table . '][' . $row['uid'] . '][delete]=1';
+                    $icon = $this->iconFactory->getIcon('actions-edit-' . $actionName, Icon::SIZE_SMALL)->render();
+                    $linkTitle = htmlspecialchars($this->getLanguageService()->getLL($actionName));
+                    $deleteAction = '<a class="btn btn-default t3js-record-delete" href="#" '
+                                    . ' data-l10parent="' . htmlspecialchars($row['l10n_parent']) . '"'
+                                    . ' data-params="' . htmlspecialchars($params) . '" data-title="' . htmlspecialchars($title) . '"'
+                                    . ' data-message="' . htmlspecialchars($warningText) . '" title="' . $linkTitle . '"'
+                                    . '>' . $icon . '</a>';
                 }
             } else {
                 $deleteAction = $this->spaceIcon;
@@ -1081,7 +1123,7 @@ class RecordListUtility extends DatabaseRecordList
                 '</a>';
         } else {
             $lockAction = $this->spaceIcon;
-        }
+            }
         $this->addActionToCellGroup($cells, $lockAction, 'lock');
 
 
@@ -1094,17 +1136,13 @@ class RecordListUtility extends DatabaseRecordList
     }
 
     /**
-     * create link for the hide/unhide and delete icon.
-     * not using tce_db.php, because we need to manipulate user's permission
+     * Returns the query builder for given table
      *
-     * @param string $params param with command (hide/unhide, delete) and records id
-     * @param string $requestURI redirect link, after process the command
-     * @return string jumpTo URL link with redirect
+     * @param string $table
+     * @return QueryBuilder
      */
-    public function actionOnClick($params, $requestURI = '')
+    protected function getQueryBuilderForTable($table)
     {
-        $redirect = '&redirect=' . ($requestURI == -1 ? "'+T3_THIS_LOCATION+'" : rawurlencode($requestURI ? $requestURI : GeneralUtility::getIndpEnv('REQUEST_URI'))) .
-            '&vC=' . rawurlencode($this->getBackendUserAuthentication()->veriCode()) . '&prErr=1&uPT=1';
-        return BackendUtility::getModuleUrl($GLOBALS['MCONF']['name']) . $params . $redirect;
+        return GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
     }
 }
